Move supporting types into subpackage
